Re-organize with opaque event structs
