+++ src/lib.rs
@@ -21,15 +21,12 @@
 //! [`run`]: run()
 //! [`Window`]: Window
 //! [`EventStream`]: EventStream
-mod event;
 mod event_stream;
 mod run;
 mod window;
 
-pub use self::event::{
-    ElementState, Event, GamepadAxis, GamepadButton, GamepadEvent, GamepadId, Key, Modifiers,
-    MouseButton, MouseScrollDelta, Pointer,
-};
+pub mod event;
+
 pub use self::event_stream::EventStream;
 pub use self::run::run;
 pub use self::window::{CursorIcon, Settings, Window};
